requirements.txt
django
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
websockets
daphne
django-cors-headers
ocpp